Download to /download and output to output.txt
Edited the fetch script to make the download/ directory and download
the data there.

I also edited parse.sh to input from the download folder and output
to output.txt and edited the error messages in parse.pl which
show example commands for running parse.pl for the sake of
consistancy.

hgnc_gene_names/output.txt was added to .gitignore as well.