pypi/parse.rb: Formatting of the abstract and switch to general output format
This now uses the format: "Package description: the package description
goes here." The first letter of the abstract is lowercased unless the
second letter was originally capitlalized too, signaling an acronym.

It's using the general output format instead of the programming
format now too. Thanks to ezgraphs for including both formats in the
code!