New fathead: HTTP Status codes
This plugin aims to provide a description for each HTTP status code.

For example, the request "HTTP 200" will output
--
200 OK
Standard response for successful HTTP requests. The actual response will
depend on the request method used. In a GET request, the response will
contain an entity corresponding to the requested resource. In a POST request
the response will contain an entity describing or containing the result of the action.
--

All the data has been scraped from the wikipedia page
https://en.wikipedia.org/wiki/List_of_HTTP_status_codes

Signed-off-by: Balthazar Rouberol <rouberol.b@gmail.com>