Save output.txt with sorted keys
This makes it easier (for humans) to read the file and find the headers.

Note: The headers are only sorted within each section, which at least
makes it consistent accross regenerations of the file. That is,
executing parse.pl a thousand times will now produce the same output (in
the same order).