Perl6Doc: Only pick up <pre> as prototype, if it is before the first <p>
Several methods (e.g.: Parameter.parcel) do not provide a prototype
before the first <p> in the HTML docs, but provide a code example
after the first <p>. Historically, this was picked up as
prototype. But as that is not a prototype, but a code example, we stop
picking it up as prototype.