Airports: Add guards for table layout
The fetched data might at some point in the future switch table
columns around, which would make it hard to see which column should go
where in the parser. hence, we add guards for the column headers. On
the one hand, this documents the expected columns. On the other, it
help detect that the format changed.