Replace cheerio with xpath and xmldom
The XML is always well formed, so cheerio feels too heavy (in terms of
DSL and dependencies) to use for parsing.

xpath has a simple API, and neither library have dependencies.

output.txt is *almost* unchanged after this commit, except for three
entries where the previous implementation missed a <desc> for some
reason.