Add additional redirects for events
This assumes that methods in the category of “events” are named after
events.  That isn’t always true, but adding extra redirects should be
harmless.

Redirects are of the form “on <event>” and “<event> event”.  Mouse
events and camelCase methods also have redirects using a split version
of the name, e.g., “on mouse down” and “mouse down event”.