Remove selector example from article titles
Searches on these titles aren’t triggering properly, even with the exact
title.  This may be because these titles use parens, brackets and
non-word characters which are being misinterpreted by the back-end, or
make them unmatchable.

Each article includes usage examples of selectors in the body, so having
them in the titles as well shouldn’t be neccessary.