Refactor redirect.py for python fathead.
 * Change, so that all articles are added to the output before
considering any redirects or related fields, because, we prioritize
articles over redirects and disambiguations. Before, there was a
possibility, that a package redirects would add a redirect with the
key of an article that appeared later in the input data from parse,
that thus, would not be added to the output, because the key was already
present. This also makes sure that package articles exist in the output
before creating related entries.